Add files via upload
This is the resource files related to the project "IoT based Garbage Monitoring System and Notification Application. The repository includes the source code, project report, snapshots and research paper done on the project. The project is basically based on the monitoring of garbage using the Arduino UNO and ESP32 module along with various kind of sensors like Ultrasonic sensor, Alcohol sensor, Smoke sensor. This uses Blynk Application to monitor the garbage through web and mobile application and it also uses the WhatsApp Bot to send messages to the user.